Added missing bson methods and removed duplication on mongodb
